files/requirements.txt
requests==2.31.0
pymysql==1.1.1
python-dotenv==1.0.0
cryptography==41.0.7   # required by pymysql for SSL connections
